try to improve performance

+++ deloldstations.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 echo "Delete data where we have not data on the station location"
-mysql  -e 'select *                                                                  from stationlocation where lt is NULL ' ognrange
-mysql  -e 'DELETE FROM positions_mgrs   WHERE station in    (select distinct station from stationlocation where lt is NULL) ' ognrange
+ps=$(mysql  -e 'select count(*)                                                                  from stationlocation where lt is NULL ' ognrange)
+nc=$(echo $ps | sed 's/^count(\*) //g')
+if [ $nc != '0' ]
+then 
+   mysql  -e 'DELETE FROM positions_mgrs   WHERE station in    (select distinct station from stationlocation where lt is NULL) ' ognrange
+fi
 mysql  -e 'DELETE FROM availability     WHERE station_id in (select distinct station from stationlocation where lt is NULL) ' ognrange
 mysql  -e 'DELETE FROM availability_log WHERE station_id in (select distinct station from stationlocation where lt is NULL) ' ognrange
 mysql  -e 'DELETE FROM roughcoverage    WHERE station in    (select distinct station from stationlocation where lt is NULL) ' ognrange
